Move CollectionLanding styled component to own file
