+++ frontend/src/metabase/collections/containers/CollectionSidebar/CollectionSidebar.styled.js
@@ -1,6 +1,8 @@
 import styled from "styled-components";
 import { Box } from "grid-styled";
 
+import { SIDEBAR_WIDTH } from "metabase/collections/constants";
+
 export const Sidebar = styled(Box.withComponent("aside"))`
   bottom: 0;
   display: flex;
@@ -11,7 +13,7 @@ export const Sidebar = styled(Box.withComponent("aside"))`
   padding-top: 32px;
   position: fixed;
   top: 65px;
-  width: ${props => props.width};
+  width: ${SIDEBAR_WIDTH};
 
   @media screen and (max-width: 768px) {
     width: 50px;
