+++ frontend/src/metabase/components/CollectionLanding/CollectionLanding.jsx
@@ -1,24 +1,12 @@
 /* eslint-disable react/prop-types */
 import React from "react";
-import { Box } from "grid-styled";
-import styled from "styled-components";
-
-import CollectionContent from "metabase/collections/containers/CollectionContent";
-import CollectionSidebar from "metabase/collections/containers/CollectionSidebar/CollectionSidebar";
 
 import * as Urls from "metabase/lib/urls";
 
 import { PageWrapper } from "metabase/collections/components/Layout";
-
-const sidebarWidth = "340px";
-
-const ContentBox = styled(Box)`
-  background-color: white;
-  height: 100%;
-  margin-left: ${sidebarWidth};
-  overflow-y: auto;
-  padding-bottom: 64px;
-`;
+import CollectionContent from "metabase/collections/containers/CollectionContent";
+import CollectionSidebar from "metabase/collections/containers/CollectionSidebar/CollectionSidebar";
+import { ContentBox } from "./CollectionLanding.styled";
 
 const CollectionLanding = ({ params: { slug }, children }) => {
   const collectionId = Urls.extractCollectionId(slug);
@@ -26,11 +14,7 @@ const CollectionLanding = ({ params: { slug }, children }) => {
 
   return (
     <PageWrapper>
-      <CollectionSidebar
-        isRoot={isRoot}
-        collectionId={collectionId}
-        width={sidebarWidth}
-      />
+      <CollectionSidebar isRoot={isRoot} collectionId={collectionId} />
       {/* For now I'm wrapping this here so that we could potentially reuse CollectionContent without
         having the specific page margin and layout concerns, TBD whether that's a good idea or needed
         */}
